requirements.txt
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.3
chromedriver-py==138.0.7204.94
colorama==0.4.6
execnet==2.1.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
pytest-xdist==3.8.0
python-dotenv==1.1.0
requests==2.32.4
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
